requirements.txt

streamlit>=1.28.0
pandas>=2.0.0
pyodbc>=4.0.35
python-dotenv>=1.0.0
requests>=2.31.0
openpyxl>=3.1.0
xlrd>=2.0.0